+++ pkg/cmd/pulumi/import.go
@@ -271,6 +271,7 @@ func newImportCmd() *cobra.Command {
 	var skipPreview bool
 	var suppressOutputs bool
 	var suppressPermaLink bool
+	var optInNonSaasPermaLink bool
 	var yes bool
 	var protectResources bool
 
@@ -391,6 +392,16 @@ func newImportCmd() *cobra.Command {
 				Type:              displayType,
 				EventLogPath:      eventLogPath,
 				Debug:             debug,
+			}
+
+			filestateBackend, err := isFilestateBackend(opts.Display)
+			if err != nil {
+				return result.FromError(err)
+			}
+			// by default, we are going to suppress the permalink when using non-SaaS backends
+			// this can be re-enabled using the `opt-in` flag
+			if filestateBackend && !optInNonSaasPermaLink {
+				opts.Display.SuppressPermaLink = true
 			}
 
 			// Fetch the project.
@@ -544,6 +555,9 @@ func newImportCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVar(
 		&suppressPermaLink, "suppress-permalink", false,
 		"Suppress display of the state permalink")
+	cmd.PersistentFlags().BoolVar(
+		&optInNonSaasPermaLink, "non-saas-permalink-enabled", false,
+		"Opt-in to showing permalinks to non-SaaS backends e.g. S3, Azure blob, GCS Storage etc.")
 	cmd.PersistentFlags().BoolVarP(
 		&yes, "yes", "y", false,
 		"Automatically approve and perform the refresh after previewing it")